fix Rails autoload issue, version bump
