+++ lib/cap-deploy-tagger/version.rb
@@ -1,3 +1,3 @@
 module CapDeployTagger
-  VERSION = "0.0.2"
+  VERSION = "0.1.0"
 end
